Orfox: add BroadcastReceiver to receive Tor status from Orbot
Run in thread so Tor status updates will be received while the Gecko event
sync is blocking the main thread. This might not be the best approach
since it probably means that the main UI will be frozen waiting for Tor to
reach ON status.

Queue URL Intents events when Orbot is not yet started

Instead of failing when opening a URL and Tor is not ready, queue those
Intents, then send them once we get STATUS_ON from Orbot.

Signed-off-by: Amogh Pradeep <amoghbl1@gmail.com>