Bug 1473509 - store principal information with the URIs to avoid having to locate documents after the fact, r=mccr8
The way we currently save resource URIs that we discovered by walking the document
isn't guaranteed to be in-order with when we save the documents they came from.
To deduce the correct principal to use to load these subresources, we need to
store it with the resource in the map of URIs we work with.

MozReview-Commit-ID: GNlNuS6TuVV

--HG--
extra : rebase_source : 99982f91971423730aa8661cf66c2dd14e276780